Revert "first pass at converting SafeMarkdown to TypeScript"
This reverts commit f7d7e31849f7ea20e304e4240a07de7f2a652a9a.
