--- apps/src/types/redactable-markdown.d.ts
@@ -1,5 +1,0 @@
-import {Plugin} from 'unified';
-declare module '@code-dot-org/redactable-markdown' {
-  const redactableMarkdown: {getParser: () => Plugin};
-  export default redactableMarkdown;
-}
