--- apps/src/types/rehype-react.d.ts
@@ -1,5 +1,0 @@
-import {Plugin} from 'unified';
-declare module 'rehype-react' {
-  const rehypeReact: Plugin;
-  export default rehypeReact;
-}
